@@ -8,11 +8,12 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.AspNetCore.Localization;
-using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.AspNetCore.SpaServices.AngularCli;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.FileProviders;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using SIL.XForge.Configuration;
@@ -28,7 +29,7 @@ namespace SIL.XForge.Scripture
 
     public class Startup
     {
-        public Startup(IConfiguration configuration, IHostingEnvironment env, ILoggerFactory loggerFactory)
+        public Startup(IConfiguration configuration, IWebHostEnvironment env, ILoggerFactory loggerFactory)
         {
             Configuration = configuration;
             Environment = env;
@@ -36,7 +37,7 @@ namespace SIL.XForge.Scripture
         }
 
         public IConfiguration Configuration { get; }
-        public IHostingEnvironment Environment { get; }
+        public IWebHostEnvironment Environment { get; }
         public ILoggerFactory LoggerFactory { get; }
         public IContainer ApplicationContainer { get; private set; }
 
@@ -68,9 +69,16 @@ namespace SIL.XForge.Scripture
         // This method gets called by the runtime. Use this method to add services to the container.
         public IServiceProvider ConfigureServices(IServiceCollection services)
         {
-            services.AddExceptionReporting(Configuration);
-
             var containerBuilder = new ContainerBuilder();
+
+            // this is a workaround for an issue in EdjCase JsonRpc.Router
+            // see https://github.com/edjCase/JsonRpc/issues/69
+            services.Configure<KestrelServerOptions>(options =>
+            {
+                options.AllowSynchronousIO = true;
+            });
+
+            services.AddExceptionReporting(Configuration);
 
             services.AddConfiguration(Configuration);
 
@@ -101,7 +109,8 @@ namespace SIL.XForge.Scripture
             services.AddLocalization(options => options.ResourcesPath = "Resources");
 
             services.AddMvc()
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_1)
+                // TODO: check if JSON.NET is required
+                .AddNewtonsoftJson()
                 .AddViewLocalization()
                 .AddDataAnnotationsLocalization(options =>
                 {
@@ -129,7 +138,7 @@ namespace SIL.XForge.Scripture
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IApplicationLifetime appLifetime, IExceptionHandler exceptionHandler)
+        public void Configure(IApplicationBuilder app, IHostApplicationLifetime appLifetime, IExceptionHandler exceptionHandler)
         {
             if (IsDevelopment)
             {
@@ -166,20 +175,24 @@ namespace SIL.XForge.Scripture
             if (SpaDevServerStartup == SpaDevServerStartup.None)
                 app.UseSpaStaticFiles();
 
+            app.UseRouting();
+
             app.UseAuthentication();
+            app.UseAuthorization();
 
             app.UseSFJsonRpc();
-
-            app.UseMvc(routes =>
-            {
-                routes.MapRoute(name: "default", template: "{controller}/{action=Index}/{id?}");
-            });
 
             app.UseRealtimeServer();
 
             app.UseMachine();
 
             app.UseSFDataAccess();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+                endpoints.MapRazorPages();
+            });
 
             // setup all server-side routes before SPA client-side routes, so that the server-side routes supercede the
             // client-side routes
